requirements.txt
Django==3.1.13
psycopg2-binary==2.9.2
djangorestframework==3.13.1
drf-spectacular==0.21.1
django-cors-headers==3.10.1
djangorestframework-simplejwt==5.0.0
pandas==1.3.5
Pillow==8.4.0
django-filter==21.1
gunicorn==20.0.4
jsonschema==4.3.2
djangorestframework-gis==0.17
django-leaflet==0.27.1
social-auth-app-django==5.0.0
redis==4.1.2
celery==5.2.3
